Warn if no steps run during epoch (#210)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/210

This adds a warning if no steps were completed during the epoch. This is to help users in case.

This may happen if users pass in an empty dataloader, or if a user unknowingly restores their dataloader state from a previously exhausted dataloader.

Since there isn't one way of pre-determining the dataloader's length without iterating over all the elements, we add a warning to the logs to make debugging this case a little bit easier.

Reviewed By: daniellepintz

Differential Revision: D39913910

fbshipit-source-id: cc437be23fc95297c9720463d4d25fe832d3a166