Always reset entry point (#246)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/246

Previously this was hardcoded whenever we called `train` / `evaluate` etc when the state was created. now that the state is created externally and passed in, it introduces the risk that these fall out of sync.

we can either set the field directly and override what was there previously, or raise an exception if an unexpected value is set

Reviewed By: daniellepintz

Differential Revision: D40527386

fbshipit-source-id: 055691056bd3d52ba89341921b03f64a6199cc6a