Add an indicator to state for the currently active phase
Summary: We have hooks such as `on_exception` which are common to train/eval/predict. in these instances, especially when using `fit` where both `train` and `eval` are called, it's difficult to know what was the currently running phase without an explicit indicator. this diff adds that indicator, which is only meaningful while the loop is being executed.

Differential Revision: D40527985

fbshipit-source-id: 2772df147b28b424df8f9df933f1d7cbcc0996dd