Add an active phase entry to the state (#245)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/245

We have hooks such as `on_exception` which are common to train/eval/predict. in these instances, especially when using `fit` where both `train` and `eval` are called, it's difficult to know what was the currently running phase without an explicit indicator. this diff adds that indicator, which is only meaningful while the loop is being executed.

Reviewed By: daniellepintz

Differential Revision: D40528003

fbshipit-source-id: e529e9854b0d5c01f71057cccc993ebfd70345db