Add data prefetching to AutoUnit
Summary:
This runs the host to device data copy in a parallel cuda stream for a performance improvement.
In order to enable this we need to pass the data iterator to the train_step rather than passing the batch.
Also removed `is_last_batch` from PhaseState and `_is_last_batch_in_epoch` from utils

Differential Revision: https://internalfb.com/D44750106

fbshipit-source-id: b50137a992460af4f52ed03dbd819c2dd1772137