Add data prefetching to AutoUnit (#366)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/366

This runs the host to device data copy in a parallel cuda stream for a performance improvement.
In order to enable this we need to pass the data iterator to the train_step rather than passing the batch.
Also removed `_is_last_batch_in_epoch` from utils since it is no longer used

Reviewed By: ananthsub

Differential Revision: D44750106

fbshipit-source-id: b3a3a505b2d740b0cd7d2d1bdc378d6291954770