Add data prefetching to AutoUnit (#366)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/366

This runs the host to device data copy in a parallel cuda stream for a performance improvement.
In order to enable this we need to pass the data iterator to the train_step rather than passing the batch.
Also removed `_is_last_batch_in_epoch` from utils since it is no longer used

Reviewed By: ananthsub

Differential Revision: D44750106

fbshipit-source-id: 6b012d330a1d2af7c0f36fb5d08cfe1f9d70797f