Add sync batch norm setting for ddp strategy (#410)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/410

As title, apply sync batchnorm conversion if using DDP: https://pytorch.org/docs/stable/generated/torch.nn.SyncBatchNorm.html
If this is not desirable, users can opt out by setting `sync_batchnorm=False`.
This changes makes converting to sync batchnorm the default, but if this is too aggressive of a change, we can make this opt-in. WDYT?

Reviewed By: daniellepintz

Differential Revision: D46450757

fbshipit-source-id: a68aba3fd48ff8c9bc595f4d695f256693dd2b6c