Flatten AutoUnit train step
Summary: This diff inlines the `_forward_and_backward` and `_run_optimizer_lr_scheduler_step` directly into the `train_step` method. This is done in order to improve the readability of the code to follow the exact order of calls. A side effect is that this disallows users from overriding those two private methods for customization, but since those are not part of the public API, it is not a concern for this diff.

Reviewed By: daniellepintz

Differential Revision: D46548404

fbshipit-source-id: b1a7180e31e5650bf80ce8b0387216a3865d038a