Account for fit in GarbageCollector callback
Summary:
Currently in the GarbageCollector callback we are not accounting for fit.

In the case of fitting we should not call `gc.disable` and `gc.enable` in `on_eval_start` and `on_eval_end` respectively.

Also, in the case of fitting we should add the total train steps completed and total eval steps completed and use that to calculate when to call `gc.collect()` based on `step_interval`.

Differential Revision: D47046140

fbshipit-source-id: b47f043117363b1576c1a586caa1c672404eb27e