Account for fit in GarbageCollector callback (#436)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/436

Currently in the GarbageCollector callback we are not accounting for fit.

In the case of fitting we should not call `gc.disable` and `gc.enable` in `on_eval_start` and `on_eval_end` respectively.

Also, in the case of fitting we should add the total train steps completed and total eval steps completed and use that to calculate when to call `gc.collect()` based on `step_interval`.

Reviewed By: crassirostris

Differential Revision: D47046140

fbshipit-source-id: d972b030597a8db23781a85be6406e133702632f