Remove auto-timing of hooks + add timing for logic in callbacks (#472)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/472

Users are now expected to time their own code using `state.timer.time()` or `get_timing_context`.

We time the code in the AutoUnit and the built-in callbacks for the user. This gives the user a lot of flexibility to time what they want to time, and makes the timer summaries more informative.

Reviewed By: ananthsub

Differential Revision: D47523530

fbshipit-source-id: a85a9b095451d4d4f1332b9d380e49401f269e47