Remove auto-timing of hooks + add timing for logic in callbacks
Summary:
Users are now expected to time their own code using `state.timer.time()` or `get_timing_context`.

We time the code in the AutoUnit and the built-in callbacks for the user. This gives the user a lot of flexibility to time what they want to time, and makes the timer summaries more informative.

Differential Revision: https://internalfb.com/D47523530

fbshipit-source-id: 585b0ed22b792239261cffecbade1f15864e526c