Add get_latest_checkpoint_path to torchsnapshot saver (#478)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/478

Add a static method to the torchsnapshot saver to find the latest checkpoint subdirectory from an input directory. This logic operates purely off the naming convention this callback uses to save snapshot directories (`epoch_N_step_M`), though this is never exposed to the caller. This could be extended for tighter validation by checking if the snapshot metadata is available inside of each subdirectory, or by sorting based on creation time, but this suffices as a simple starting point.

For navigating remote filesystems, this uses the `fsspec` library

Callers would use this function as following:
```
latest_snapshot_path = TorchSnapshotSaver.get_latest_checkpoint_path(dirpath)
if latest_snapshot_path:
    TorchSnapshotSaver.restore(latest_snapshot_path, unit, ...)
```

Reviewed By: JKSenthil

Differential Revision: D47748160

fbshipit-source-id: 2f3a366c374d24dfc26523ce7ec8b1bd5c727cdf