Add always-on iteration timer to framework (#439)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/439

Add two counters, containing the iteration and time blocked on data for the last iteration. These are stored and recorded using a timer that does not synchronise with CUDA, and that only records these two values.

- iteration time is recorded in the training loop for all jobs.
- data time is recorded if the training loop does the data fetching, otherwise the user needs to instrument the logic which reads the data from the iterable.

Had to rework the approach slightly since the state does not seem the right place. With the changes, it is more natural to make it similar to progress, and store the info in a Stateful. This has the additional advantage that all values should be storable and be restored with the checkpoints.

Also rather than the last value, the last LOWER_BOUND values are stored. I hardcoded this in the torchtnt state, to not bloat parameters, since intuitively the last 1e4 values for each timer action here should be enough for monitoring purposes.

Reviewed By: daniellepintz, ananthsub

Differential Revision: D46853794

fbshipit-source-id: 28043ec16cbc23f9e75e851328898e40b9f73bda