In TorchSnapshotSaver initialize optimizer state before restoring (#554)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/554

Optimizer state is lazily initialized, meaning when you create an optimizer and call state_dict() it will have fewer keys than when you create an optimizer, call optimizer.step() and then call state_dict().

If you create an optimizer and then try to restore the optimizer from a previously saved state_dict() (which was saved after optimizer.step()), it will fail because not all of the keys are in the state_dict() of the initial optimizer.

The workaround is to initialize optimizer state with `optimizer.step()` before restoring.

Reviewed By: ananthsub, janeyx99

Differential Revision: D49563335

fbshipit-source-id: 55e591888e0b891ec69aed2e70578bf041cd8835