Check for snapshot metadata for path to restore from (#559)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/559

With async saving, some writes can succeed before an error occurs, causing the checkpoint directory to be created, but the snapshot metadata won't be written.

The exception is caught and logged to not kill the main process, as control has returned back to the user.

On restoration, currently the code checks for only the existence and naming convention of the directory, even though the directory doesn't contain the essential snapshot metadata needed to restore.

This change adds that check to avoid passing a directory path that doesn't have the metadata to restore from.

Reviewed By: RdoubleA

Differential Revision: D49789432

fbshipit-source-id: 0bd0334fda9402a8423a774c1691d654ba6c330c