replace del w/ pop to remove from app_state
Summary:
# Context
It is assumed that app_state will have train_progress and eval_progress keys. But this may not be the case if the user manually overrides the `app_state`. This will cause `KeyError: 'train_progress'`when loading checkpoint
# This diff
Replaces `del app_state[KEY]` to `app_state.pop(KEY, None)` which will delete the key if it exists, or do nothing otherwise

Reviewed By: galrotem

Differential Revision: D50034777