replace del w/ pop to remove from app_state (#570)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/570

# Context
It is assumed that app_state will have train_progress and eval_progress keys. But this may not be the case if the user manually overrides the `app_state`. This will cause `KeyError: 'train_progress'`when loading checkpoint
# This diff
Replaces `del app_state[KEY]` to `app_state.pop(KEY, None)` which will delete the key if it exists, or do nothing otherwise

Reviewed By: galrotem

Differential Revision: D50034777

fbshipit-source-id: 2fe3e07621a9464efd59016b314e2eef570410da