add no-op strategy (#578)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/578

# Context
Currently there is no way for user to skip `prepare_module` call in AutoUnit constructor. This will always augment the model in some way, depending on what's passed as the `strategy` (and if `None` is passed, module is moved to device). This may not be desirable for users who want to do all their processing of their module on their own and don't want even the module to be moved to device.

# This Diff
Adds a no-op strategy for users that essentially skips `prepare_modules` application of strategy on the module, and avoids moving module to device.

# Future
Once this unblocks immediate users, we can maybe look to introduce a `MoveToDevice` strategy and let `None` be the equivalent of the no-op strategy

Reviewed By: galrotem

Differential Revision: D50202998

fbshipit-source-id: 6a2eb71595337d2e5a206f73f761fd9c1993376f