separate build and publish doc scripts (#666)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/666

# Context
Publish docs part of the script is failing since the directory generated by the build script isn't shared to the publish one
 {F1238220367}

# This Diff
Creates a separate script for building and publishing the static docs. Building is only run on pull requests, build + publishing script is run on commit / land.

bypass-github-export-checks

Reviewed By: galrotem

Differential Revision: D52525534

fbshipit-source-id: 76e972767fcff490c666b8a91e1643b07f057237