emphasize generation cleanup in gc callback
Summary:
In the GC callback, it mentions generation 1 garbage collection occurs every step. It's not clear what what kind of garbage collection the `step_interval` is responsible for. 

This diff changes docstring to emphasize the generation 1 cleanup on each step, and that the `step_interval` is in charge of the most significant garbage collection generation.

Differential Revision: D52675919