torchtnt: limit number of max bins for parameter monitor to 1M
Summary: As described in [this post](https://fburl.com/workplace/p4183k43) we could end up with a stuck job if parameter monitor is enabled for a large model training job. Instead of removing the callback from such jobs, we could also reduce the number of bins to improve efficiency and potentially avoid stuck jobs.

Differential Revision: D53064677