Reduce `_generate_checkpoint_and_upkeep` code complexity (#783)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/783

The function _generate_checkpoint_and_upkeep is important for the checkpointing logic. But it has a linter warning indicating that it is too complex, and a there is a TODO to extract some logic into a separate function.

Let's do a small refactor to improve readability and reduce function complexity, but avoid breaking changes or regressions.

**Potential Future Changes**

Note that while doing this change, I found two small bugs that we can fix. They are documented in this Bento notebook: https://fburl.com/anp/gqoezved I did not fix any of them here to avoid having a refactor + logic changes.

Additionally, there is this user request that we can handle in this function. Again it was not modified here but we can decide what to do and change later: https://fb.workplace.com/groups/cu.training.framework.users/permalink/1147131179616886/

Reviewed By: galrotem

Differential Revision: D55881050

fbshipit-source-id: 46d0777a2dc5208763628fecdac8c12a5573e407