[DCP][TNT] Adds check for FSPDOptimizer wrapper
DCP expects the entire optimizer to be initialized in the state dict before calling load. Unfortunately, since optimizers are sometimes lazy loaded, this means dcp_saver has to do a check for optimizer objects and ensure they are properly initialized.

This diff adds a check for optimizers which are hidden inside wrappers. An alternative to this would be to implement in the FSDP wrapper in under utils/prepare_module, but I found that change to be higher risk, and then we'd also have the change in both places.

Differential Revision: [D56075363](https://our.internmc.facebook.com/intern/diff/D56075363/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D56075363/)!

ghstack-source-id: 222855240
Pull Request resolved: https://github.com/pytorch/tnt/pull/788