remove <= 2.0 version checks
Summary:
# Context
Part of effort to make TorchTNT OSS compliant.

TorchTNT uses latest pytorch apis. Sometimes these are only available in nightlies. When that's the case, users in OSS using a stable pytorch release will see import errors when they use TorchTNT. This can be guarded against by importing conditionally, only when the compatible pytorch version is detected in the environment.

We have plenty of these checks already. However, at the same time we don't want to bload TNT with too many of these version checks everywhere. Currently we have various version checks for Pytorch 1.0.

I propose to make Pytorch 2.0+ a hard dependency for TorchTNT going forward. This will
* remove existing version check bloat in TNT
* force users to use latest features from pytorch
And Pytorch 2.0 is documented to be 100% backwards compatible, so no bugs or errors should show up

# This Diff

Removes all the Pytorch version checks below 2.0 in various places of the codebase, and all the helper functions which check for version

Reviewed By: galrotem

Differential Revision: D56446353

fbshipit-source-id: 2a594e21fa755f249f0f40e352aa90e4476d83ca