Configurable planner and storage writer in the dcp saver save API (#821)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/821

Configurable planner and storage writer in the DCP saver save API.

# This Stack
DCP saver is the TorchTNT callback which allows checkpointing via the Distributed Checkpointing APIs. Current implementation doesn't expose the Save Planner and Storage Writer in the API for clients to plug in their implementations. It enforces the default planner and FsspecWriter.

# This diff
- DCP save and async save APIs now support planner and storage writer allowing clients to plug in their implementations.
- Introduces a knob option to plug in storage writer component with storage efficiency optimizations

Reviewed By: JKSenthil

Differential Revision: D56921724

fbshipit-source-id: b60c34c6df38e02c0af9a1db4ffd243d382fd621