Revert D64379847: Multisect successfully blamed "D64379847: [AIInfra][DCP] Fix the load state dict in case of partial loading" for one test failure (#931)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/931

This diff reverts D64379847
D64379847: [AIInfra][DCP] Fix the load state dict in case of partial loading by saumishr causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_py_predictor_apps_mgp_non_hero_ocr_ocr_pipeline_test#main](https://www.internalfb.com/intern/test/844425109859400/)

Here's the Multisect link:
https://www.internalfb.com/multisect/12282180
Here are the tasks that are relevant to this breakage:
T198471974: 10+ tests failing for mgenai_ml_infra

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: saumishr

Differential Revision: D64448222

fbshipit-source-id: b1a3f244d38067010a4dae4b61029ef2ce9bf4a7