Use TNT's ManifoldPathHandler for listing checkpoints internally
Summary:
We've faced multiple issues in the past where users register incompatible implementations of Manifold path handlers to fsspec, causing errors when listing and even loading the checkpoints.

Currently [one user is facing an error because of this](https://fb.workplace.com/groups/277527419809135/permalink/1625534775008386/), and it's tricky to debug because the error is not reproducible outside of that particular project (because of the specific dependencies being used)

Most internal customers should be using storage optimizations and then use modelstore components within DCP APIs, but we still use fsspec for listing latest and best checkpoints.

We need to make sure that our own implementation is used to list the checkpoints. Note we can't modify directly on checkpoint utils because it's OSS, while filesystem is internal.

Differential Revision: D65370757