Support buffer size argument in `broadcast_str` util (#988)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/988

This commit adds a new test case to the DistributedTest class to test the
broadcast_str function with a fixed buffer size. The test case checks that
the broadcasted value is correct when the fixed buffer size is larger than,
equal to, and smaller than the length of the input string.
Note: This commit also includes some minor changes to the existing test cases
to make them more robust.
Changes:
- Added new test case test_broadcast_str_fixed_buffer_size to DistributedTest
- Updated existing test cases to use spawn_multi_process instead of spawn
Error:
The test case is currently failing due to an "enforce fail" error in the Gloo
backend. Further investigation is needed to determine the root cause of this
error.

Reviewed By: JKSenthil

Differential Revision: D72077879