Calculate _step_requires_iterator only once (#1025)
Summary:
Pull Request resolved: https://github.com/pytorch/tnt/pull/1025

Minor optimisation opportunity that I noticed when poking around in here.

`_step_requires_iterator` uses the `inspect` module to look at type hints.  This is somewhat expensive (tens of microseconds), so you probably don't want to recalculate it on every batch.  It should be constant, so we can calculate it once during init.

Alternatively, we could stick a `funcutils.cache` decorator on `_step_requires_iterator`.  That has its pros and cons, LMK if you think it's better.

Reviewed By: galrotem

Differential Revision: D80776986

fbshipit-source-id: 0410671dfee94d063a084c8a2c5e5cccf9741768